+++ integration/downstreamHttp/downstreamreadtimeout_test.go
@@ -1,4 +1,4 @@
-package downstream
+package downstreamHttp
 
 import (
 	"fmt"
